# openapi.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local Development"
    },
    {
      "url": "https://staging-api.yourdomain.com/api/v1",
      "description": "Staging Environment"
    },
    {
      "url": "https://api.yourdomain.com/api/v1",
      "description": "Production Environment"
    }
  ],
  "info": {
    "title": "Omni-Backend API",
    "version": "1.0.0",
    "description": "Generated by nestia - https://github.com/samchon/nestia",
    "license": {
      "name": "MIT"
    }
  },
  "components": {
    "schemas": {
...
      "ConversationListItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "externalId": {
            "type": "string"
          },
          "platform": {
            "type": "string",
            "enum": [
              "X",
              "Instagram",
              "LinkedIn",
              "Telegram",
              "Email"
            ]
          },
          "accountId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "unreadCount": {
            "type": "number"
          },
          "state": {
            "type": "string",
            "enum": [
              "open",
              "closed",
              "archived"
            ]
          },
          "text": {
            "type": "string"
          },
          "online": {
            "type": "boolean"
          },
          "lastSeen": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastMessageStatus": {
            "type": "string",
            "enum": [
              "sent",
              "delivered",
              "read",
              "failed"
            ]
          },
          "conversationType": {
            "type": "string",
            "enum": [
              "parent",
              "child"
            ]
          },
          "receiverEmail": {
            "type": "string"
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConversationListItem"
            }
          },
          "parentConversationId": {
            "type": "string"
          },
          "subject": {
            "type": "string"
          },
          "out": {
            "type": "boolean"
          },
          "sentAt": {
            "type": "string",
            "format": "date-time"
          },
          "direction": {
            "type": "string"
          },
          "messageStatus": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "sentBy": {
            "$ref": "#/components/schemas/PartialUserDTO"
          },
          "provideOriginalPayload": {
            "$ref": "#/components/schemas/object.o2"
          },
          "externalMessageId": {
            "type": "string"
          },
          "threadId": {
            "type": "string"
          },
          "messageId": {
            "type": "string"
          },
          "inReplyTo": {
            "type": "string"
          },
          "references": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "participants": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "bccRecipients": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "externalId",
          "platform",
          "accountId",
          "name",
          "unreadCount",
          "createdAt",
          "updatedAt"
        ]
      },
      "PartialUserDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "role": {
            "$ref": "#/components/schemas/RoleDTO"
          },
          "platformAccess": {
            "type": "array",
            "items": {}
          },
          "phoneNumber": {
            "type": "string"
          },
          "disabled": {
            "type": "boolean"
          },
          "emailVerifiedAt": {
            "type": "string",
            "format": "date-time"
          },
          "twoFactorSecret": {
            "type": "string"
          }
        },
        "required": [],
        "description": "Make all properties in T optional"
      },
      "RoleDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "createdAt",
          "updatedAt"
        ]
      },
      "object.o2": {
        "type": "object",
        "properties": {},
        "required": []
      },
...
      "WipeEmailInboxBody": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string"
          },
          "folderPath": {
            "type": "string"
          }
        },
        "required": [
          "accountId"
        ]
      },
...
    }
  },
  "paths": {
...
    "/messages/conversations": {
      "get": {
        "tags": [
          "Messages"
        ],
        "parameters": [
          {
            "name": "platform",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "X",
                "Instagram",
                "LinkedIn",
                "Telegram",
                "Email"
              ]
            },
            "required": false
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "number"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ConversationListItem"
                      }
                    },
                    "nextCursor": {
                      "type": "string",
                      "nullable": true
                    },
                    "hasMore": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "data",
                    "nextCursor",
                    "hasMore"
                  ]
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "messages",
          "conversations",
          "fetchConversations"
        ],
        "x-samchon-controller": "MessageController"
      }
    },
    "/messages/conversations/all": {
      "get": {
        "tags": [
          "Messages"
        ],
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "number"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ConversationListItem"
                      }
                    },
                    "nextCursor": {
                      "type": "string",
                      "nullable": true
                    },
                    "hasMore": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "data",
                    "nextCursor",
                    "hasMore"
                  ]
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "messages",
          "conversations",
          "all",
          "fetchAllConversations"
        ],
        "x-samchon-controller": "MessageController"
      }
    },
    "/messages": {
      "get": {
        "tags": [
          "Messages"
        ],
        "parameters": [
          {
            "name": "conversationId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "name": "threadId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "externalMessageId": {
                        "type": "string"
                      },
                      "direction": {
                        "type": "string",
                        "enum": [
                          "inbound",
                          "outbound"
                        ]
                      },
                      "subject": {
                        "type": "string",
                        "nullable": true
                      },
                      "status": {
                        "type": "string",
                        "enum": [
                          "sent",
                          "delivered",
                          "read",
                          "failed"
                        ]
                      },
                      "role": {
                        "type": "string",
                        "enum": [
                          "owner",
                          "pa"
                        ]
                      },
                      "text": {
                        "type": "string",
                        "nullable": true
                      },
                      "out": {
                        "type": "boolean",
                        "nullable": true
                      },
                      "sentAt": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                      },
                      "threadId": {
                        "type": "string",
                        "nullable": true
                      },
                      "messageId": {
                        "type": "string",
                        "nullable": true
                      },
                      "inReplyTo": {
                        "type": "string",
                        "nullable": true
                      },
                      "references": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        },
                        "nullable": true
                      },
                      "sentBy": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "firstName": {
                            "type": "string"
                          },
                          "lastName": {
                            "type": "string"
                          },
                          "email": {
                            "type": "string"
                          },
                          "role": {
                            "type": "object",
                            "properties": {
                              "name": {
                                "type": "string",
                                "enum": [
                                  "PA",
                                  "super-admin"
                                ]
                              }
                            },
                            "required": [
                              "name"
                            ]
                          }
                        },
                        "required": [
                          "id",
                          "firstName",
                          "lastName",
                          "email",
                          "role"
                        ],
                        "nullable": true
                      },
                      "provideOriginalPayload": {
                        "$ref": "#/components/schemas/object.o2"
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "participants": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    },
                    "required": [
                      "id",
                      "externalMessageId",
                      "direction",
                      "status",
                      "role",
                      "sentBy",
                      "provideOriginalPayload",
                      "createdAt",
                      "updatedAt",
                      "participants"
                    ]
                  }
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "messages",
          "fetchMessages"
        ],
        "x-samchon-controller": "MessageController"
      }
    },
    "/messages/email/cleanup": {
      "post": {
        "tags": [
          "Messages"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WipeEmailInboxBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "deletedCount": {
                      "type": "number"
                    },
                    "accountId": {
                      "type": "string"
                    },
                    "folderPath": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "deletedCount",
                    "accountId",
                    "folderPath"
                  ]
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "messages",
          "email",
          "cleanup",
          "wipeEmailInbox"
        ],
        "x-samchon-controller": "MessageController"
      }
    },
...
  },
...
}